Import chromium-browser_50.0.2661.102.orig.tar.xz from Ubuntu Yakkety
All .gitignore files dropped, as this would have prevented much of the
tarball being imported.
 # find . -name .gitignore -delete

Removed files over 50MB in size to make GitHub happy.
 # find . -type f -size +49M -delete

+++ chrome/browser/chromeos/external_metrics.cc
@@ -4,6 +4,8 @@
 
 #include "chrome/browser/chromeos/external_metrics.h"
 
+#include <stddef.h>
+
 #include <map>
 #include <string>
 
@@ -29,7 +31,7 @@ namespace {
 bool CheckValues(const std::string& name,
                  int minimum,
                  int maximum,
-                 size_t bucket_count) {
+                 uint32_t bucket_count) {
   if (!base::Histogram::InspectConstructionArguments(
       name, &minimum, &maximum, &bucket_count))
     return false;
