+++ chrome/browser/safe_browsing/incident_reporting/platform_state_store_win_unittest.cc
@@ -4,8 +4,7 @@
 
 #include "chrome/browser/safe_browsing/incident_reporting/platform_state_store.h"
 
-#include "base/prefs/pref_notifier_impl.h"
-#include "base/prefs/testing_pref_store.h"
+#include "base/macros.h"
 #include "base/strings/utf_string_conversions.h"
 #include "base/test/test_reg_util_win.h"
 #include "base/test/test_simple_task_runner.h"
@@ -16,6 +15,8 @@
 #include "chrome/test/base/testing_profile.h"
 #include "chrome/test/base/testing_profile_manager.h"
 #include "components/pref_registry/pref_registry_syncable.h"
+#include "components/prefs/pref_notifier_impl.h"
+#include "components/prefs/testing_pref_store.h"
 #include "components/syncable_prefs/testing_pref_service_syncable.h"
 #include "testing/gtest/include/gtest/gtest.h"
 
@@ -99,7 +100,7 @@ class PlatformStateStoreWinTest : public ::testing::Test {
     ASSERT_EQ(ERROR_SUCCESS,
               key.ReadValue(base::UTF8ToUTF16(kProfileName_).c_str(),
                             &buffer[0], &data_size, &data_type));
-    EXPECT_EQ(REG_BINARY, data_type);
+    EXPECT_EQ(static_cast<DWORD>(REG_BINARY), data_type);
     ASSERT_EQ(kTestDataSize_, data_size);
     EXPECT_EQ(std::string(&buffer[0], data_size),
               std::string(&kTestData_[0], kTestDataSize_));
