+++ chrome/installer/util/delete_reg_value_work_item_unittest.cc
@@ -4,6 +4,7 @@
 
 #include <windows.h>
 
+#include "base/macros.h"
 #include "base/memory/scoped_ptr.h"
 #include "base/strings/string16.h"
 #include "base/test/test_reg_util_win.h"
@@ -92,8 +93,8 @@ TEST_F(DeleteRegValueWorkItemTest, DeleteExistingValue) {
   DWORD type = 0;
   DWORD size = 0;
   EXPECT_EQ(ERROR_SUCCESS, key.ReadValue(kNameEmpty, NULL, &size, &type));
-  EXPECT_EQ(REG_SZ, type);
-  EXPECT_EQ(0, size);
+  EXPECT_EQ(static_cast<DWORD>(REG_SZ), type);
+  EXPECT_EQ(0u, size);
 }
 
 // Try deleting a value that doesn't exist.
