+++ chrome_elf/blacklist/test/blacklist_test.cc
@@ -2,10 +2,13 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+#include <stddef.h>
+
 #include "base/environment.h"
 #include "base/files/file_path.h"
 #include "base/files/scoped_temp_dir.h"
 #include "base/i18n/case_conversion.h"
+#include "base/macros.h"
 #include "base/path_service.h"
 #include "base/scoped_native_library.h"
 #include "base/strings/string16.h"
@@ -65,7 +68,7 @@ class BlacklistTest : public testing::Test {
     base::FilePath current_dir;
     ASSERT_TRUE(PathService::Get(base::DIR_EXE, &current_dir));
 
-    for (int i = 0; i < arraysize(test_data); ++i) {
+    for (size_t i = 0; i < arraysize(test_data); ++i) {
       // Ensure that the dll has not been loaded both by inspecting the handle
       // returned by LoadLibrary and by looking for an environment variable that
       // is set when the DLL's entry point is called.
@@ -233,7 +236,7 @@ TEST_F(BlacklistTest, LoadBlacklistedLibrary) {
   EXPECT_EQ(num_initially_blocked_, num_blocked_dlls);
 
   // Add all DLLs to the blacklist then check they are blocked.
-  for (int i = 0; i < arraysize(test_data); ++i) {
+  for (size_t i = 0; i < arraysize(test_data); ++i) {
     EXPECT_TRUE(TestDll_AddDllToBlacklist(test_data[i].dll_name));
   }
   CheckBlacklistedDllsNotLoaded();
@@ -254,7 +257,7 @@ TEST_F(BlacklistTest, AddDllsFromRegistryToBlacklist) {
       HKEY_CURRENT_USER,
       blacklist::kRegistryFinchListPath,
       KEY_QUERY_VALUE | KEY_SET_VALUE);
-  for (int i = 0; i < arraysize(test_data); ++i) {
+  for (size_t i = 0; i < arraysize(test_data); ++i) {
     finch_blacklist_registry_key.WriteValue(test_data[i].dll_name,
                                             test_data[i].dll_name);
   }
